Support for Avorion Module
Adds support for avorion.
Known issues:

RCON/input does not work, need to investigate method of setting rcon password in server.ini file for this.

Server process does not respond to any close options. Only method to shutdown server is to kill or use commands in game and this can lead to loss of galaxy data.